+++ NEWS.md
@@ -1,3 +1,10 @@
+# fozziejoin 0.0.5
+
+- [TODO] Factor support now included
+- Reverted a change from v0.0.4 wherein speed distance calculation methods differ by operating system (Windows vs. everything else). The supposed speed gains were actually flaws in the evaluation. Reverted back to a single method for all OS's.
+- README documentation was updated to fix an error in the example usage code. `fuzzyjoin` was a required import for the `misspellings` dataset.
+
+
 # fozziejoin 0.0.4
 
 - Performance improvements: